requirements.txt
flask
requests
flask-sqlalchemy
flask-login
pandas
pycryptodome
